pyproject.toml
[project]
name = "trucker-log"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.1",
    "django-rest-knox>=5.0.2",
    "djangorestframework>=3.16.1",
]
